gralloc: Fix PMEM heap allocation for 3.0 kernels
This is needed for PMEM camera libraries.

Symptoms: gralloc gives an error saying invalid argument to map /dev/pmem_adsp, and dmesg says that ION is trying to mmap it. This issue is likely being caused by the ION drivers in the 3.0 kernel as this issue does not exist with 3.4 kernels.

Signed-off-by: Sultanxda <sultanxda@gmail.com>